backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
joblib==1.3.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
reportlab==4.0.7
